test_cli/requirements.txt
astroid==2.11.7
autopep8==2.0.1
dill==0.3.4
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
mimesis==4.1.3
mypy==0.971
mypy-extensions==0.4.3
platformdirs==2.4.0
pycodestyle==2.10.0
pydash==5.1.2
pylint==2.13.9
python-dotenv==0.20.0
tomli==1.2.3
typed-ast==1.5.4
typing-extensions==4.1.1
wrapt==1.14.1
pandas
numpy
scikit-learn
Faker
